Fix all actionable CodeQL security and quality alerts
Security:
- BootReceiver: verify intent action before processing (improper intent verification)
- CredentialManager: remove cleartext SharedPreferences fallback; throw on encryption failure

Errors:
- PlusActivity: remove dead QkTextView branch (QkTextView extends TextView, branch unreachable)

Warnings:
- GalleryPagerAdapter: remove useless ?. safe-calls on non-null ExoPlayer
- ComposeActivity: remove redundant targetPosition bounds checks (already guarded by totalItemCount > 0)
- BackupRepositoryImpl: guard error log with messageCount > 0 to eliminate trivially-false path
- HSVPickerView: add explicit parentheses to clarify hue calculation operator precedence
- SearchAdapter: add explicit parentheses around > comparisons before ==
- QkEditText: add explicit parentheses around bitwise-and before != 0
- RewardedAdManager: widen multiplication to Long before int overflow can occur
- Experiment: replace new Random().nextInt() with idiomatic kotlin .random()

Notes:
- WebSyncRepositoryImpl: remove unread loginResponse local variable

+++ presentation/src/main/java/com/charles/messenger/common/util/RewardedAdManager.kt
@@ -145,7 +145,7 @@ class RewardedAdManager @Inject constructor(
         // Calculate ad-free time based on points
         // 1 point = 30 minutes
         val adFreeMinutes = points * 30
-        val adFreeMillis = adFreeMinutes * 60 * 1000L
+        val adFreeMillis = adFreeMinutes.toLong() * 60L * 1000L
 
         // Add to existing ad-free time if already active
         val currentAdFreeEndTime = prefs.adFreeEndTime.get()
